added draft of extrusion example

+++ polychrom/hdf5_format.py
@@ -119,29 +119,37 @@ class HDF5Reporter(object):
     def __init__(self, folder, max_data_length=50, 
                  h5py_dset_opts={"compression":"gzip"}, 
                  overwrite=False, 
+                 blocks_only=False
                 ):
         """
         
         """
+        prefixes = ["blocks", "applied_forces", "initArgs", "starting_conformation", "energy_minimization"]
         if not os.path.exists(folder):
             os.mkdir(folder)
             
         if overwrite: 
             for the_file in os.listdir(folder):
                 file_path = os.path.join(folder, the_file)
-                if os.path.isfile(file_path):
-                    os.remove(file_path)
+                if os.path.isfile(file_path):                    
+                    for prefix in prefixes:
+                        if the_file.startswith(prefix):
+                            os.remove(file_path)
                 else:
                     raise IOError("Subfolder in traj folder; not deleting. Ensure folder is correct and delete manually. ")
                         
                     
         if len(os.listdir(folder)) != 0:
-            raise RuntimeError(f"folder {folder} is not empty")
+            for the_file in os.listdir(folder):            
+                for prefix in prefixes:
+                    if the_file.startswith(prefix):
+                        raise RuntimeError(f"folder {folder} is not empty")
         self.counter = {}
         self.datas = {}
         self.max_data_length = max_data_length
         self.h5py_dset_opts = h5py_dset_opts
         self.folder = folder
+        self.blocks_only = blocks_only
         
 
     def report(self, name, values):
@@ -149,9 +157,10 @@ class HDF5Reporter(object):
         
         
         if name not in ["data"]:
-            filename = f"{name}_{count}.h5"
-            with h5py.File(os.path.join(self.folder,filename)) as file: 
-                _write_group(values, file, dset_opts=self.h5py_dset_opts)
+            if not self.blocks_only:
+                filename = f"{name}_{count}.h5"
+                with h5py.File(os.path.join(self.folder,filename)) as file: 
+                    _write_group(values, file, dset_opts=self.h5py_dset_opts)
                 
         else:
             self.datas[count] = values
